backend/pyproject.toml
[project]
name = "infra-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn>=0.30.0",
    "docker>=7.1.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.9",
    "boto3>=1.34.0",
    "qdrant-client>=1.9.0",
    "pymongo>=4.6.0",
]
